+++ src/main/java/dk/sjd/demo/Model/Repositories/EmployeeRepository.java
@@ -24,7 +24,7 @@ public class EmployeeRepository implements IEmployeeRepository {
 
             SqlRowSet sqlRowSet1 = jdbc.queryForRowSet("SELECT sum(hours) AS totalHours FROM shifts WHERE name='" + sqlRowSet.getString("name") + "'");
             while(sqlRowSet1.next()){
-                employees.add(new Employee(sqlRowSet.getInt("id"), sqlRowSet.getString("name"), sqlRowSet1.getInt("totalHours")));
+                employees.add(new Employee(sqlRowSet.getInt("id"), sqlRowSet.getString("name"), sqlRowSet1.getDouble("totalHours")));
             }
         }
 
@@ -47,7 +47,7 @@ public class EmployeeRepository implements IEmployeeRepository {
         SqlRowSet sqlRowSet = jdbc.queryForRowSet("SELECT * FROM employees WHERE id =" + id + "");
 
         if (sqlRowSet.next()){
-            return new Employee(sqlRowSet.getInt("id"), sqlRowSet.getString("name"), sqlRowSet.getInt("totalHours"));
+            return new Employee(sqlRowSet.getInt("id"), sqlRowSet.getString("name"), sqlRowSet.getDouble("totalHours"));
         }
         return new Employee();
     }
